rename package from pyptoject to spark_rapids_pytools
