+++ user_tools/setup.cfg
@@ -1,6 +1,6 @@
 [options.entry_points]
 console_scripts = 
     spark_rapids_dataproc = spark_rapids_dataproc_tools.dataproc_wrapper:main
-    spark_rapids_user_tools = pyrapids.wrapper:main
+    spark_rapids_user_tools = spark_rapids_pytools.wrapper:main
 [options.package_data]
 * = *.json, *.yaml
