+++ user_tools/src/spark_rapids_pytools/common/exceptions.py
