+++ user_tools/src/spark_rapids_pytools/common/prop_manager.py
