+++ user_tools/src/spark_rapids_pytools/resources/emr-configs.json
